+++ crates/matrix-sdk-ui/CHANGELOG.md
@@ -21,6 +21,9 @@ All notable changes to this project will be documented in this file.
 
 ### Features
 
+- [**breaking**] Extend `TimelineFocus::Event` to allow marking the target
+  event as the root of a thread.
+  ([#6050](https://github.com/matrix-org/matrix-rust-sdk/pull/6050))
 - [**breaking**] Remove `TimelineEventTypeFilter` which has been replaced by
   the more generic `TimelineEventFilter`.
   ([#6070](https://github.com/matrix-org/matrix-rust-sdk/pull/6070/))